Only reflect in the reflecting object's own color
Previously we were doing perfect (white) reflection but coloured
diffusion, which looks intuitively weird.

This is an attempt to only reflect in the same spectrum as the object's
own color.

To make this a bit clearer I've moderated the colors of the objects so
they have at least some of each channel in them; otherwise the
completely disjoint colors of the three main objects leads to weird
voids (which I think is correct, just looks surprising).

I've used the term "albedo" for the components of reflection, which is
effectively a unit vector in the color space. This may or may not be the
right word.