Only render background color if ray intersects with nothing
Having the background color returned from recursive trace calls has
problematic effects; in particular, it acts like a light source on
reflective objects, but doesn't interact with diffuse objects at all.

This alters our trace() function to return null both when max recursion
depth is reached, or if no intersection is found. In the outermost
trace() call we then handle this case with the background color, and in
recursive reflection calls we use black instead.