Fix Camera#trace()
Our Camera#trace() function should return a ray with an origin at the
eye, and direction pointing from the eye to the "pixel" on the film
being rendered.

However, the eye doesn't factor into our direction calculation at the
moment, and we're just returning the pixel point in 3D space as the
direction. This isn't easily apparent as our eye is so close to the
origin, but makes quite a visual difference once corrected.