Add "Lambertian" material
This models the surface scatter approach described in [1], in which we
basically bounce the ray off in any random direction from the incident
surface.

This is *supposedly* modelling a Lambertian (i.e. perfectly diffuse)
surface, but since we're only currently using scattered rays in our
reflection component, the results are terrible and this is pretty
obviously Not Quite Right. As a result we're not actually applying it to
any surface just yet.

[1] http://viclw17.github.io/2018/07/30/raytracing-reflecting-materials/